+++ src/app/features/tenant/tenant-detail/tenant-detail.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from "@angular/router";
 import {TenantService} from "../services/tenant.service";
+import {TenantBusinessModel} from "../../../common/business-models/tenant.business-model";
 
 @Component({
   selector: 'app-tenant-detail',
@@ -8,9 +9,10 @@ import {TenantService} from "../services/tenant.service";
   styleUrls: ['./tenant-detail.component.scss']
 })
 export class TenantDetailComponent implements OnInit {
-
+  public tenant?: TenantBusinessModel;
   public tenantID!: string;
   public isLoading = false;
+  public subtitle = "Locataire";
 
   constructor(
     private activatedRoute: ActivatedRoute,
@@ -21,6 +23,18 @@ export class TenantDetailComponent implements OnInit {
 
   ngOnInit(): void {
     this.tenantID = this.activatedRoute.snapshot.params['id_tenant'];
+    this.getTenant();
+  }
+
+  private getTenant(): void {
+    this.tenantService.getTenant(this.tenantID).subscribe({
+      next: (tenant: TenantBusinessModel) => {
+        this.tenant = tenant;
+      },
+      error: () => {
+        // TODO handle error
+      }
+    });
   }
 
   public delete(): void {