+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Nomenclature :
 - Updated toolbar : when user is not authenticated the links and the menu doesn't exist
 - Updated register : added button to go on register page from login and added button to go on login page from register -
   removed temporary links on menu burger
+- Added display of tenant and owner detail
 
 ### [0.1.1] - 2023-03-02